Completed 3.3 (Moving Ball with Player)
I first finished off the physics materials from 3.2 and attached them to the ball and player. However a dynamic friction of 6 for the Rough material stopped my player from moving, so I left it at 4 which feels nice. Then I modified BallController.cs to move the ball with the player as it moves.