Completed Section 4.1 (Fall Trigger / Tags)
Added Fall Trigger logic for the pins, triggers the debug log 10 times for a strike and accurately counts how many pins were knocked down in all other outcomes. I also tried to add the Trigger / Tag logic for the gutters, so that the gutters wouldn't shoot pins out, not sure if it was correct but I tried!