Completed Section 4.2 (Score)
I added logic for the scoring system in the game using GameManager. I then added a text element in the in-game UI to display the score to the user. I am having some issues where sometimes the pins physics get a little wonky and they don't hit the ground right away, or get hung up on each other and the FallTrigger never triggers, so I increased the width of the FallTrigger Box Collider to make it slightly more accurate.