Completed Section 4.3 (Reset)
All of the reset logic seems to be working. Fixed that one pin falling through the ground again, and dramatically increased the FallTrigger width for detecting if pins have fallen down. Pins were frequently laying on top of others so that their hitbox was upwards on an angle and not making contact with the ground. So I widened it so that if a pin is more than roughly 45 degrees towards the ground the isPinFallen boolean should trigger, giving more accurate scores.